Add comprehensive test suite and complete Android build configuration
- Added 40+ unit tests covering core functionality:
  * AuthRepositoryTest: 13 tests for user registration and login
  * BuscaParcaApiTest: 17 tests for API data models and validation
  * LocationUtilsTest: 9 tests for distance calculations and formatting
  * ParkingMLModelTest: 14 tests for ML prediction algorithm
  * DataModelsTest: 2 tests for data model conversions
  * DatabaseTest: 12 instrumented tests for Room database operations

- Enhanced build configuration:
  * Added Mockito, Mockito-Kotlin, Coroutines Test dependencies
  * Added Room Testing and Architecture Components Testing libraries
  * Migrated from KAPT to KSP for better Kotlin compatibility
  * Configured JDK 17 build environment for Android Gradle Plugin 8.0.2

- Fixed critical bugs:
  * String.format() now uses Locale.US for consistent number formatting
  * Fixed ParkingZone hex color literals with .toInt() conversion
  * Fixed LocationTrackingService location type conversions (Double to Float)
  * Added missing iconBackground color resource

- Added project resources:
  * Launcher icons for all density buckets (mdpi to xxxhdpi)
  * ProGuard rules configuration
  * Gradle wrapper files for reproducible builds
  * BUILD_STATUS.md with comprehensive build documentation
  * test-android-build.sh verification script

All tests passing successfully. APK builds at 7.8 MB.

+++ android-app/app/src/main/java/com/buscaparca/data/models/ParkingZone.kt
@@ -27,9 +27,9 @@ data class ParkingZone(
     
     fun getHotZoneColor(): Int {
         return when {
-            successRate >= 0.7 -> 0x8000FF00 // Green with transparency
-            successRate >= 0.4 -> 0x80FFFF00 // Yellow with transparency
-            else -> 0x80FF0000 // Red with transparency
+            successRate >= 0.7 -> 0x8000FF00.toInt() // Green with transparency
+            successRate >= 0.4 -> 0x80FFFF00.toInt() // Yellow with transparency
+            else -> 0x80FF0000.toInt() // Red with transparency
         }
     }
 }
